requirements.txt
langchain-core~=0.2.0
langchain~=0.2.0
langchain-community~=0.2.0
langchain-openai~=0.1.25
langchain_huggingface~=0.0.3
langchain-chroma~=0.1.0

langgraph~=0.0.69
chromadb~=0.5.0
openai~=1.40.0
sentence-transformers~=3.0.0
huggingface-hub~=0.23.0
transformers~=4.41.0
tiktoken~=0.7.0
python-dotenv~=1.0.0
requests
httpx~=0.27.0

torch~=2.2.2

pandas~=2.2.3
PyPDF2